printsiginfo: decode si_syscall field using syscall_name_arch
It was subtly broken all this time as si_arch may be not equivalent
to strace's current personality's audit arch: this may well happen when
siginfo_t is decoded as an argument (like in ptrace(PTRACE_SETSIGINFO)),
as the author of the comment in v4.21~43 should have definitely known.

* src/printsiginfo.c: Remove "nr_prefix.c" include.
[HAVE_SIGINFO_T_SI_SYSCALL] (print_si_syscall): Remove.
(print_si_info): Use PRINT_SYSCALL_NAME for si_syscall field printing.
* tests/ptrace.c: Update expected output, add checks for the cases when
{CUR,PERS0,M32,MX32}_AUDIT_ARCH are defined.
* NEWS: Mention it.